* master: (25 commits)
  fix typo
  add comment
  fix Coverity warning: Division or modulo by zero
  fix Coverity warning: Dereference before null check
  fix Coverity warning: Dereference before null check
  fix Coverity warning: Explicit null dereferenced
  fix Coverity warning: Dereference after null check
  fix Coverity warning: Explicit null dereferenced
  fix Coverity warning: Uninitialized scalar field
  fix Coverity warning: Uninitialized scalar field
  fix Coverity warning: Dereference after null check
  fix Coverity warning: Dereference null return value
  fix Coverity warning: Unintended sign extension
  fix Coverity warning: Missing break in switch
  fix Coverity warning: Misuse of memcmp-style function
  fix Coverity warning: Dereference after null check
  fix Coverity warning: Dereference after null check
  fix Coverity warning: Dereference before null check
  fix Coverity warning: Uninitialized scalar field
  fix Coverity warning: Null pointer dereferences
  ...

+++ src/base/Sysutils.cpp
@@ -1023,7 +1023,7 @@ static DWORD FindMatchingFile(TSearchRec & Rec)
   WORD Hi = (Rec.Time & 0xFFFF0000) >> 16;
   WORD Lo = Rec.Time & 0xFFFF;
   FileTimeToDosDateTime(reinterpret_cast<LPFILETIME>(&LocalFileTime), &Hi, &Lo);
-  Rec.Time = (Hi << 16) + Lo;
+  Rec.Time = ((Integer)Hi << 16) + Lo;
   Rec.Size = Rec.FindData.nFileSizeLow || static_cast<Int64>(Rec.FindData.nFileSizeHigh) << 32;
   Rec.Attr = Rec.FindData.dwFileAttributes;
   Rec.Name = Rec.FindData.cFileName;
